Move Perspective logic to sdk
