+++ frontend/packages/console-dynamic-plugin-sdk/src/index.ts
@@ -4,3 +4,5 @@ export * from './api/common-types';
 
 // Extension types
 export * from './extensions';
+
+export * from './perspective';
