+++ frontend/packages/console-dynamic-plugin-sdk/src/perspective/index.ts
@@ -1,0 +1,2 @@
+export { default as useActivePerspective } from './useActivePerspective';
+export * from './perspective-context';
