+++ frontend/packages/console-dynamic-plugin-sdk/src/perspective/perspective-context.ts
@@ -1,10 +1,9 @@
 import * as React from 'react';
-
-export type PerspectiveType = string;
+import { PerspectiveType } from '../extensions';
 
 export type PerspectiveContextType = {
   activePerspective?: PerspectiveType;
   setActivePerspective?: React.Dispatch<React.SetStateAction<PerspectiveType>>;
 };
 
-export const PerpsectiveContext = React.createContext<PerspectiveContextType>({});
+export const PerspectiveContext = React.createContext<PerspectiveContextType>({});
