+++ frontend/packages/console-shared/src/hooks/index.ts
@@ -23,7 +23,7 @@ export * from './useUserSettings';
 export * from './useUserSettingsCompatibility';
 export * from './hpa-hooks';
 export * from './usePinnedResources';
-export * from './useActivePerspective';
+export * from './perspectiveUtils';
 export * from './useActiveNamespace';
 export * from './useIsMobile';
 export * from './useResizeObserver';
