+++ frontend/packages/console-shared/src/hooks/perspectiveUtils.ts
@@ -1,0 +1,7 @@
+import { PerspectiveType } from '@console/dynamic-plugin-sdk';
+import { USERSETTINGS_PREFIX } from '../constants';
+
+const PERSPECTIVE_VISITED_FEATURE_KEY = 'perspective.visited';
+
+export const getPerspectiveVisitedKey = (perspective: PerspectiveType): string =>
+  `${USERSETTINGS_PREFIX}.${PERSPECTIVE_VISITED_FEATURE_KEY}.${perspective}`;
