+++ frontend/packages/console-shared/src/hooks/usePinnedResources.ts
@@ -1,10 +1,9 @@
 import { useMemo, useCallback } from 'react';
 import * as _ from 'lodash';
-import { isPerspective, Perspective } from '@console/dynamic-plugin-sdk';
+import { isPerspective, Perspective, useActivePerspective } from '@console/dynamic-plugin-sdk';
 import { referenceForExtensionModel } from '@console/internal/module/k8s';
 import { useExtensions } from '@console/plugin-sdk';
 import { PINNED_RESOURCES_LOCAL_STORAGE_KEY } from '../constants';
-import { useActivePerspective } from './useActivePerspective';
 import { useTelemetry } from './useTelemetry';
 import { useUserSettingsCompatibility } from './useUserSettingsCompatibility';
