+++ frontend/packages/console-shared/src/hooks/useTabbedTableBreadcrumb.ts
@@ -5,12 +5,12 @@ import { useTranslation } from 'react-i18next';
 // @ts-ignore
 import { useSelector } from 'react-redux';
 import { match as RMatch } from 'react-router-dom';
+import { useActivePerspective } from '@console/dynamic-plugin-sdk';
 import { getBreadcrumbPath } from '@console/internal/components/utils/breadcrumbs';
 import { K8sKind } from '@console/internal/module/k8s';
 import { getActiveNamespace } from '@console/internal/reducers/ui';
 import { RootState } from '@console/internal/redux';
 import { ALL_NAMESPACES_KEY } from '../constants/common';
-import { useActivePerspective } from './useActivePerspective';
 
 type Match = RMatch<{ url: string }>;
