+++ frontend/packages/dev-console/src/components/import/jar/UploadJar.tsx
@@ -2,15 +2,16 @@ import * as React from 'react';
 import { Formik, FormikHelpers } from 'formik';
 import * as _ from 'lodash';
 import { useTranslation } from 'react-i18next';
-import { WatchK8sResultsObject, Perspective, isPerspective } from '@console/dynamic-plugin-sdk';
+import {
+  WatchK8sResultsObject,
+  Perspective,
+  isPerspective,
+  useActivePerspective,
+} from '@console/dynamic-plugin-sdk';
 import { history } from '@console/internal/components/utils';
 import { K8sResourceKind } from '@console/internal/module/k8s';
 import { useExtensions } from '@console/plugin-sdk';
-import {
-  ALL_APPLICATIONS_KEY,
-  useActivePerspective,
-  usePostFormSubmitAction,
-} from '@console/shared/src';
+import { ALL_APPLICATIONS_KEY, usePostFormSubmitAction } from '@console/shared/src';
 import { sanitizeApplicationValue } from '@console/topology/src/utils';
 import { BuilderImage } from '../../../utils/imagestream-utils';
 import { getBaseInitialValues } from '../form-initial-values';
