+++ frontend/packages/dev-console/src/components/monitoring/overview/MonitoringOverviewAlerts.tsx
@@ -2,10 +2,11 @@ import * as React from 'react';
 import { Alert } from '@patternfly/react-core';
 import * as _ from 'lodash';
 import { Link } from 'react-router-dom';
+import { useActivePerspective } from '@console/dynamic-plugin-sdk';
 import { Alert as AlertType } from '@console/internal/components/monitoring/types';
 import { labelsToParams } from '@console/internal/components/monitoring/utils';
 import { fromNow } from '@console/internal/components/utils/datetime';
-import { sortMonitoringAlerts, useActivePerspective } from '@console/shared';
+import { sortMonitoringAlerts } from '@console/shared';
 import { getAlertType } from './monitoring-overview-alerts-utils';
 import './MonitoringOverviewAlerts.scss';
