+++ frontend/packages/knative-plugin/src/components/add/brokers/AddBroker.tsx
@@ -3,11 +3,10 @@ import { Formik, FormikHelpers } from 'formik';
 import { safeLoad } from 'js-yaml';
 import { useTranslation } from 'react-i18next';
 import { handleRedirect } from '@console/dev-console/src/components/import/import-submit-utils';
-import { Perspective, isPerspective } from '@console/dynamic-plugin-sdk';
+import { Perspective, isPerspective, useActivePerspective } from '@console/dynamic-plugin-sdk';
 import { history } from '@console/internal/components/utils';
 import { K8sResourceKind, k8sCreate } from '@console/internal/module/k8s';
 import { useExtensions } from '@console/plugin-sdk';
-import { useActivePerspective } from '@console/shared';
 import { EditorType } from '@console/shared/src/components/synced-editor/editor-toggle';
 import { EventingBrokerModel } from '../../../models';
 import { AddBrokerFormYamlValues } from '../import-types';
