+++ frontend/packages/knative-plugin/src/components/add/channels/AddChannel.tsx
@@ -2,13 +2,13 @@ import * as React from 'react';
 import { Formik } from 'formik';
 import { useTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
-import { Perspective, isPerspective } from '@console/dynamic-plugin-sdk';
+import { Perspective, isPerspective, useActivePerspective } from '@console/dynamic-plugin-sdk';
 import { history } from '@console/internal/components/utils';
 import { K8sResourceKind, k8sCreate, modelFor, referenceFor } from '@console/internal/module/k8s';
 import { getActiveApplication } from '@console/internal/reducers/ui';
 import { RootState } from '@console/internal/redux';
 import { useExtensions } from '@console/plugin-sdk';
-import { ALL_APPLICATIONS_KEY, useActivePerspective } from '@console/shared';
+import { ALL_APPLICATIONS_KEY } from '@console/shared';
 import { sanitizeApplicationValue } from '@console/topology/src/utils/application-utils';
 import { getCreateChannelResource } from '../../../utils/create-channel-utils';
 import { handleRedirect } from '../../../utils/create-eventsources-utils';
