+++ frontend/public/components/nav/perspective-nav.tsx
@@ -9,8 +9,9 @@ import {
   isNavSection,
   NavSection as PluginNavSection,
   isNavItem,
-} from '@console/dynamic-plugin-sdk/src';
-import { useActivePerspective, usePinnedResources } from '@console/shared';
+  useActivePerspective,
+} from '@console/dynamic-plugin-sdk';
+import { usePinnedResources } from '@console/shared';
 import { K8sKind, modelFor, referenceForModel } from '../../module/k8s';
 import { getSortedNavItems } from './navSortUtils';
 import confirmNavUnpinModal from './confirmNavUnpinModal';
